requirements.txt
transformers
gradio
pillow
requests
numpy
torch